feat: move all interfaces to own folder
