+++ src/CitMovie.Business/ManagerInterfaces/ICountryManager.cs
@@ -1,0 +1,7 @@
+namespace CitMovie.Business;
+
+public interface ICountryManager
+{
+    Task<IEnumerable<CountryResult>> GetAllCountriesAsync(int page, int pageSize);
+    Task<int> GetTotalCountriesCountAsync();
+}
