+++ src/CitMovie.Business/ManagerInterfaces/IGenreManager.cs
@@ -1,0 +1,7 @@
+namespace CitMovie.Business;
+
+public interface IGenreManager
+{
+    Task<IEnumerable<GenreResult>> GetAllGenresAsync(int page, int pageSize);
+    Task<int> GetTotalGenresCountAsync();
+}
