+++ src/CitMovie.Business/ManagerInterfaces/IJobCategoryManager.cs
@@ -1,0 +1,7 @@
+namespace CitMovie.Business;
+
+public interface IJobCategoryManager
+{
+    Task<IEnumerable<JobCategoryResult>> GetAllJobCategoriesAsync(int page, int pageSize);
+    Task<int> GetTotalJobCategoriesCountAsync();
+}
