+++ src/CitMovie.Business/ManagerInterfaces/ILoginManager.cs
