+++ src/CitMovie.Business/ManagerInterfaces/IMediaManager.cs
