+++ src/CitMovie.Business/ManagerInterfaces/ITitleTypeManager.cs
