+++ src/CitMovie.Business/ManagerInterfaces/IUserManager.cs
