--- src/CitMovie.Business/Managers/ICountryManager.cs
@@ -1,7 +1,0 @@
-namespace CitMovie.Business;
-
-public interface ICountryManager
-{
-    Task<IEnumerable<CountryResult>> GetAllCountriesAsync(int page, int pageSize);
-    Task<int> GetTotalCountriesCountAsync();
-}
