--- src/CitMovie.Business/Managers/IGenreManager.cs
@@ -1,7 +1,0 @@
-namespace CitMovie.Business;
-
-public interface IGenreManager
-{
-    Task<IEnumerable<GenreResult>> GetAllGenresAsync(int page, int pageSize);
-    Task<int> GetTotalGenresCountAsync();
-}
