--- src/CitMovie.Business/Managers/IJobCategoryManager.cs
@@ -1,7 +1,0 @@
-namespace CitMovie.Business;
-
-public interface IJobCategoryManager
-{
-    Task<IEnumerable<JobCategoryResult>> GetAllJobCategoriesAsync(int page, int pageSize);
-    Task<int> GetTotalJobCategoriesCountAsync();
-}
